+++ shaders/include/constants.glsl
@@ -4,8 +4,9 @@
 #define VRTX_DEBUG_OFF 0
 #define VRTX_DEBUG_FINAL_PASS_MARKER 1
 #define VRTX_DEBUG_UV_GRADIENT 2
+#define VRTX_DEBUG_MATERIAL_CATEGORY 3
 
-#define VRTX_DEBUG_VIEW 0 // [0 1 2] Debug view mode
+#define VRTX_DEBUG_VIEW 0 // [0 1 2 3] Debug view mode
 
 #define VRTX_MATERIAL_UNKNOWN 0